refactor : (Auth) Facade 적용
